fun zero to NaN (quality of zeros) as optional, default False. Moved many fun from prep_steps to fun_fm.py
