+++ notes.txt
@@ -1,0 +1,5 @@
+ALERT! when finished the fitting part improvements, re-integrate the univariate tests . == DONE
+
+bug when running prepare in example2 dataset, function calculate_meanEnri in loop for met__i in metabolites
+
+
